Moved Settings class files to Models directory
